refactor(console-ui): phase 3 — split lib/api.ts + decompose streamChat (F1/F8/F9)
The 851-line god-client and its 354-line streamChat (ESLint cognitive
complexity 119) are broken up behind a barrel that preserves the @/lib/api
import path (zero caller churn):

- lib/api/{types,models,pricing,billing,keys,providers,health}.ts + index.ts
  over the shared lib/http/proxy-client. lib/api.ts deleted.
- lib/chat/{think-parser,sse,stream}.ts: streamChat is now a thin orchestrator
  over a ThinkStreamParser (the multi-format think FSM), an SSE payload iterator,
  and small transport/error/metrics helpers.
- F9: ChatMessage now imports parseThinkFromContent from the shared
  think-parser instead of carrying a second copy of the grammar.
- F8: useAuth's provisioning sequence is one `provisionApiKey` callback used by
  both the mount effect and the key-expired handler (kills the promise-nesting
  warnings).

Tests: +__tests__/think-parser.test.ts. eslint warnings 130 -> 116 (streamChat
cog-119 + handleContentToken cog-31 + useAuth promise nesting all gone).
Invariant held: build green, vitest 352 pass / 4 pre-existing fail.
